+++ CHANGELOG.md
@@ -5,10 +5,15 @@
 ### New
 
 - [#166][is_166]: Start using Postmark for emails ([#168][pr_168])
+- [#167][is_167]: Create and schedule weekly report email ([#169][pr_169])
 
 [is_166]: https://github.com/JSn1nj4/ElliotDerhay.com/issues/166
 
+[is_167]: https://github.com/JSn1nj4/ElliotDerhay.com/issues/167
+
 [pr_168]: https://github.com/JSn1nj4/ElliotDerhay.com/pull/168
+
+[pr_169]: https://github.com/JSn1nj4/ElliotDerhay.com/pull/169
 
 ## Version 2.6.4